+++ src/constants.js
@@ -32,7 +32,9 @@ export const DTYPE_VALUES = {
     format: GL.R32F,
     dataFormat: GL.RED,
     type: GL.FLOAT,
-    max: 2 ** 31 - 1,
+    // Not sure what to do about this one - a good use case for channel stats, I suppose:
+    // https://en.wikipedia.org/wiki/Single-precision_floating-point_format.
+    max: 3.4 * 10 ** 38,
     TypedArray: Float32Array
   }
 };